+++ config/routes.rb
@@ -1,4 +1,9 @@
 VboxwebRb::Application.routes.draw do |map|
+  match 'vm/imports'              => 'import#index',    :as => 'vm_imports'
+  match 'vm/imports/new'          => 'import#new',      :as => 'vm_new_import'
+  match 'vm/imports/:id/progress' => 'import#progress', :as => 'vm_import_progress'
+  match 'vm/imports/:id'          => 'import#show',     :as => 'vm_import'
+
   match 'vm/:uuid/exports'              => 'export#index',    :as => 'vm_exports'
   match 'vm/:uuid/exports/new'          => 'export#new',      :as => 'vm_new_export'
   match 'vm/:uuid/exports/:id/progress' => 'export#progress', :as => 'vm_export_progress'